requirements.txt
beautifulsoup4==4.13.3
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
et_xmlfile==2.0.0
filelock==3.18.0
Flask==3.1.0
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==5.3.1
MarkupSafe==3.0.2
mpmath==1.3.0
numpy==2.2.4
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
pillow==11.3.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.1
requests==2.32.3
six==1.17.0
soupsieve==2.6
sympy==1.13.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
Werkzeug==3.1.3
